requirements.txt
bilibili-api-python
curl_cffi
python-dotenv
qrcode
qrcode-terminal

Pillow
